+ comment

+++ 233_Number_of_Digit_One.cpp
@@ -13,17 +13,23 @@ public:
 
         size_t count = 0;
         for (int i = 1; i <= num_size; i++) {
-            int a = n / size_t(pow(10, i));
+            // 前部
+            int a = n / pow(10, i);
+            // 尾巴
             int b = n % size_t(pow(10, i - 1));
-            count += (size_t(pow(10, i - 1)) * a);
+            // 当前位固定，则有前面位数那么多的个1
+            count += pow(10, i - 1) * a;
 
+            // 当前数字
             auto digit = num_str[num_size - i] - '0';
             if (digit > 1) {
-                count += size_t(pow(10, i - 1));
+                // 后面全加
+                count += pow(10, i - 1);
             } else if (digit == 1) {
+                // 加后面,但只有1
                 count += b;
                 count += 1;
-            }
+            } // 0就没了
         }
         return count;
     }
